refactor: Adjust file structure for non-controller modules
