+++ apps/service/modules/guards/index.ts
@@ -2,3 +2,4 @@ export * from './auth.guard';
 export * from './token.service';
 export * from './permissions.guard';
 export * from './guards.module';
+export * from './app-keys.guard';
