+++ apps/service/src/app-keys/app-keys.module.ts
@@ -1,20 +1,14 @@
 import { Module } from '@nestjs/common';
 
-import { DatabaseModule } from 'src/database';
+import { DatabaseModule } from 'modules/database';
 
 import { AppKeysService } from './app-keys.service';
 import { AppReportService } from './app-report.service';
 import { TransformerService } from './transformer.service';
 import { AppKeysController } from './app-keys.controller';
-import { AppKeyGuard } from './app-keys.guard';
 import { AppKeysScheduleController } from './app-keys.schedule.controller';
 
-const PROVIDERS = [
-  AppKeysService,
-  AppReportService,
-  TransformerService,
-  AppKeyGuard,
-];
+const PROVIDERS = [AppKeysService, AppReportService, TransformerService];
 
 @Module({
   imports: [DatabaseModule],
