+++ apps/service/src/app-keys/index.ts
@@ -1,5 +1,3 @@
 export * from './app-keys.module';
 export * from './app-keys.controller';
 export * from './app-keys.service';
-export { TransformerService as AppKeyTransformerService } from './transformer.service';
-export * from './app-keys.guard';
