+++ apps/service/src/applications/index.ts
@@ -1,0 +1,3 @@
+export * from './applications.module';
+export * from './applications.controller';
+export * from './applications.service';
