+++ apps/service/src/applications/transformer.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { DataType } from '@zpanel/core';
 
-import { Model } from 'src/database';
+import { Model } from 'modules/database';
 
 // ----------
