+++ apps/service/src/auth/auth.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 
-import { DatabaseModule } from 'src/database/database.module';
+import { DatabaseModule } from 'modules/database/database.module';
 
 import { AuthService } from './auth.service';
 import { TransformerService } from './transformer.service';
