+++ apps/service/src/media/index.ts
@@ -1,4 +1,3 @@
 export * from './media.controller';
 export * from './media.module';
 export * from './media.service';
-export { TransformerService as MediaTransformerService } from './transformer.service';
