+++ apps/service/src/media/media.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 
-import { DatabaseModule } from 'src/database';
+import { DatabaseModule } from 'modules/database';
 
 import { MediaService } from './media.service';
 import { TransformerService } from './transformer.service';
