+++ apps/service/src/permissions/index.ts
@@ -1,4 +1,3 @@
 export * from './permissions.module';
 export * from './permissions.controller';
 export * from './permissions.service';
-export * from './transformer.service';
