+++ apps/service/src/projects/portfolio/index.ts
@@ -1,4 +1,3 @@
 export * from './portfolio.module';
 export * from './portfolio.controller';
 export * from './portfolio.service';
-export * from './transformer.service';
