+++ apps/service/src/projects/portfolio/portfolio.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 
-import { DatabaseModule } from 'src/database';
+import { DatabaseModule } from 'modules/database';
 
 import { PortfolioService } from './portfolio.service';
 import { TransformerService } from './transformer.service';
