+++ apps/service/src/roles/index.ts
@@ -1,4 +1,3 @@
 export * from './roles.module';
 export * from './roles.controller';
 export * from './roles.service';
-export { TransformerService as RoleTransformerService } from './transformer.service';
