+++ apps/service/src/roles/roles.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 
-import { DatabaseModule } from 'src/database';
+import { DatabaseModule } from 'modules/database';
 
 import { RolesService } from './roles.service';
 import { TransformerService } from './transformer.service';
